backend/requirements.txt
fastapi>=0.115.0,<0.120
uvicorn[standard]>=0.30.0,<0.40
python-multipart>=0.0.9
pymupdf>=1.24.10,<2.0
anthropic>=0.40.0
pydantic>=2.9.0,<3.0
python-dotenv>=1.0.0
